usr/article/doDelete 권한 체크

+++ src/main/java/com/lkj/exam/demo/controller/UsrArticleController.java
@@ -63,34 +63,66 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
 		
 		Article article = articleService.getArticle(id);
-		
+
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id), id);
+			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 게시물에 대한 권한이 없습니다.", id));
 		}
 		
 		articleService.deleteArticle(id);
 		
-		return ResultData.from("S-1", Ut.f("%d번 게시물은 삭제했습니다.", id), id);
+		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제했습니다.", id), id);
 	}
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Integer> doModify(int id, String title, String body) {
+	public ResultData<Integer> doModify(HttpSession httpSession, int id, String title, String body) {
+		
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
 		
 		Article article = articleService.getArticle(id);
-		
+
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id), id);
+			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 게시물에 대한 권한이 없습니다.", id));
 		}
 		
 		articleService.modifyArticle(id, title, body);
 		
 		article = articleService.getArticle(id);
 		
-		return ResultData.from("S-1", Ut.f("%d번 게시물은 수정했습니다.", id), id);
+		return ResultData.from("S-1", Ut.f("%d번 게시물을 수정했습니다.", id), id);
 	}
 	
 	@RequestMapping("/usr/article/getArticle")
